+++ packages/design/src/index.ts
@@ -3,3 +3,4 @@ export * from "./frick.design.js";
 export * from "./model.js";
 export * from "./resolver.js";
 export * from "./validate.js";
+export * from "./validate-styles.js";